Rework entry points to TypeScript
Fixes the dts build invocation and moves robot.js and bin/hubot.js to
TypeScript. Types and usages are lightly massaged to be correct, and
HUBOT_SCRIPTS environment variable support is nuked (evidence of whose
use is basically 0 on the internet).

`Robot` is also lightly refactored for better typing, though this
refactoring is currently incomplete and untested.