Add correct types for log library
The @types/log types are for a much newer version of the library that
has been completely reworked. For backwards compatibility with older
scripts, don't move to this version, but fix the type definitions for
the old one.

Also fix a couple of invocations that were passing possible nulls or
undefineds as the first parameter to the log call.